feat(okx): fractional-contract sizing 'B' (notional = 2x equity)
Root-fix the small-account over-leverage: the executor floored size to WHOLE
contracts (int(target_notional/per_contract)), so on $89 it was forced to 1
contract = $610 = ~7x, and needed 10x leverage just to clear the floor. OKX
actually allows 0.01-contract lots ($6 min, verified via public instruments
API), so this was self-inflicted.

New sizing: notional = NOTIONAL_LEV_MULT(2.0) x equity, rounded to the 0.01 lot
step. On $89 -> ~0.24 contracts (~$180, ~2x effective). Scales with equity (no
leverage creep). Sim (169-trade WF + injected -10% gap): 2x survives the gap,
10x = ruin.

Made the whole execution pipeline fractional-aware:
- 7 int(size_contracts) casts -> float (executor x3, approval, reconciler,
  rest, state) incl. the close-path int(0.29)=0 bug that would block closing
- Position/TradeIntent.size_contracts: int -> float
- reconciler SIZE_TOLERANCE 0.001 -> 0.005 (float dust vs real mismatch)
- DB v7_okx_positions.size_contracts INT -> DECIMAL(12,4) (live+archive applied)
- new test: $89 -> fractional ~0.24, ~2x, order submits the fractional sz

275 okx tests pass.

+++ indicator/okx/approval.py
@@ -49,7 +49,7 @@ class TradeIntent:
     stop_price: float
     atr: float
     stop_dist: float
-    size_contracts: int
+    size_contracts: float     # fractional lots allowed (OKX lotSz 0.01)
     size_frac: float
     notional_usd: float
     equity_before: float
@@ -84,7 +84,7 @@ class TradeIntent:
             stop_price=float(d["stop_price"]),
             atr=float(d["atr"]),
             stop_dist=float(d["stop_dist"]),
-            size_contracts=int(d["size_contracts"]),
+            size_contracts=float(d["size_contracts"]),
             size_frac=float(d["size_frac"]),
             notional_usd=float(d["notional_usd"]),
             equity_before=float(d["equity_before"]),
